requirements.txt
cycler==0.11.0
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pygame==2.1.2
pygame-widgets==1.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
